requirements.txt
aiogram>=3.13.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
yt-dlp>=2024.12.13
shazamio>=0.1.6.1
